alter to confirm

+++ NewWeightRecord/form.js
@@ -14,10 +14,13 @@ window.addEventListener("load",function(){
         console.log(time);
         console.log(weight);
 
-        alert("정말 기록하시겠습니까?");
-        //localStorage에 저장
-        localStorage.setItem("date",date);
-        localStorage.setItem("time",time);
-        localStorage.setItem("weight",weight);
+        var con = confirm("정말 기록하시겠습니까?");
+        if(con == true){ //localStorage에 저장
+            localStorage.setItem("date",date);
+            localStorage.setItem("time",time);
+            localStorage.setItem("weight",weight);
+            location.href = "table.html"
+        }
+
     }
 });
